Check input files exist before submitting to SLURM
For some reason, condor-ce and BLAH can get out of sync -- the net
result is that BLAH is told to submit jobs whose input files (and even
parent directories!) don't exist in condor's spol directory. These jobs
are doomed to fail and be later restarted, causing a lot of unnecessary
churn on the SLURM scheduler.

Before submitting jobs to SLURM, verify that the input files exist. If
they don't, error out early.